Breaking circular imports: Move base classes to yubico_base.py
This breaks the yubikey <-> yubikey_usb_hid circular import
